Layout taking a mobile-first approach, use media queries and negative margins on the <img>, and apply a border to the table rows
Having a mobile-first approach in mind I started
to layout each section and add spacing between
the  elements using margin.

The <img> in smaller viewports take the full width
of the recipe container while in bigger viewports
it respects the padding of the recipe container.
In order to do this I also took a mobile-first
approach. First I applied the following rules to
the recipe container (the <article
class="recipe">):
 -padding: 0 1.625rem 2.5rem;
Then using the image selector I set these rules:
 - width: calc(100% + (1.625rem * 2));
 - margin-left: -1.625rem

To make the <img> take the full width of its
container in bigger viewports I use a media query
and set the following rules:
- .recipe {
    padding: 2.8rem;
}
- .recipe-img {
    width: 100%;
    margin-left: 0;
}